+++ browser/brave_profile_prefs.cc
@@ -53,7 +53,7 @@
 #include "brave/components/search_engines/brave_prepopulated_engines.h"
 #include "brave/components/speedreader/common/buildflags/buildflags.h"
 #include "brave/components/tor/buildflags/buildflags.h"
-#include "brave/components/webcompat_reporter/browser/webcompat_reporter_utils.h"
+#include "brave/components/webcompat_reporter/common/pref_names.h"
 #include "build/build_config.h"
 #include "chrome/browser/prefetch/pref_names.h"
 #include "chrome/browser/prefs/session_startup_pref.h"
@@ -493,7 +493,7 @@ void RegisterProfilePrefs(user_prefs::PrefRegistrySyncable* registry) {
   registry->SetDefaultPrefValue(prefs::kSearchSuggestEnabled,
                                 base::Value(false));
 
-  webcompat_reporter::RegisterProfilePrefs(registry);
+  webcompat_reporter::prefs::RegisterProfilePrefs(registry);
 }
 
 }  // namespace brave
