+++ browser/webcompat_reporter/sources.gni
@@ -15,6 +15,7 @@ brave_browser_webcompat_reporter_deps = [
   "//brave/browser:browser_process",
   "//brave/common",
   "//brave/components/webcompat_reporter/browser",
+  "//brave/components/webcompat_reporter/common",
   "//brave/components/webcompat_reporter/common:mojom",
   "//components/keyed_service/content",
   "//mojo/public/cpp/bindings",
