+++ components/webcompat_reporter/browser/webcompat_reporter_utils.cc
@@ -10,8 +10,6 @@
 #include "base/containers/contains.h"
 #include "base/containers/fixed_flat_set.h"
 #include "brave/components/brave_shields/core/common/brave_shield_constants.h"
-#include "brave/components/webcompat_reporter/common/pref_names.h"
-#include "components/prefs/pref_registry_simple.h"
 
 namespace {
 
@@ -36,11 +34,6 @@ namespace webcompat_reporter {
 
 bool NeedsToGetComponentInfo(std::string_view component_id) {
   return base::Contains(kComponentIds, component_id);
-}
-
-void RegisterProfilePrefs(PrefRegistrySimple* registry) {
-  registry->RegisterBooleanPref(prefs::kContactInfoSaveFlagPrefs, true);
-  registry->RegisterStringPref(prefs::kContactInfoPrefs, "");
 }
 
 std::string BoolToString(bool value) {
