+++ components/webcompat_reporter/common/BUILD.gn
@@ -6,7 +6,12 @@
 import("//mojo/public/tools/bindings/mojom.gni")
 
 static_library("common") {
-  sources = [ "pref_names.h" ]
+  sources = [
+    "pref_names.cc",
+    "pref_names.h",
+  ]
+
+  deps = [ "//components/prefs" ]
 }
 
 mojom("mojom") {
