+++ ios/browser/api/webcompat_reporter/BUILD.gn
@@ -33,6 +33,7 @@ source_set("webcompat_reporter") {
     "//base",
     "//brave/base/mac",
     "//brave/components/webcompat_reporter/browser:browser",
+    "//brave/components/webcompat_reporter/common",
     "//brave/ios/browser/keyed_service",
     "//components/component_updater:component_updater",
     "//components/keyed_service/core",
