+++ ios/browser/api/webcompat_reporter/webcompat_reporter_service_factory.mm
@@ -6,7 +6,7 @@
 #include "brave/ios/browser/api/webcompat_reporter/webcompat_reporter_service_factory.h"
 
 #include "brave/components/webcompat_reporter/browser/webcompat_reporter_service.h"
-#include "brave/components/webcompat_reporter/browser/webcompat_reporter_utils.h"
+#include "brave/components/webcompat_reporter/common/pref_names.h"
 #include "brave/ios/browser/api/webcompat_reporter/webcompat_reporter_service_delegate.h"
 #include "components/keyed_service/ios/browser_state_dependency_manager.h"
 #include "components/pref_registry/pref_registry_syncable.h"
@@ -42,7 +42,7 @@ WebcompatReporterServiceFactory::~WebcompatReporterServiceFactory() {}
 
 void WebcompatReporterServiceFactory::RegisterBrowserStatePrefs(
     user_prefs::PrefRegistrySyncable* registry) {
-  webcompat_reporter::RegisterProfilePrefs(
+  webcompat_reporter::prefs::RegisterProfilePrefs(
       static_cast<PrefRegistrySimple*>(registry));
 }
